lambda_code/requirements.txt
kafka-python==2.0.2
boto3
